+++ screens/NewStock.tsx
@@ -1,10 +1,11 @@
-import { Text, View, StatusBar,Button, FlatList, StyleSheet } from 'react-native';
+import { Text, View, StatusBar,Button, FlatList, StyleSheet, Alert } from 'react-native';
 import { useDatabase } from '../utils/DatabaseCotext';
 import GroupButton from '../components/GroupButton';
 import { useEffect, useState } from 'react';
 import ModalStock from '../components/ModalStock';
 import TableItem from '../components/TableItem';
 import SmallButton from '../components/SmallButton';
+import ImgButton from '../components/ImgButton';
 
 
 export default function NewStock({navigation}) {
@@ -21,7 +22,7 @@ export default function NewStock({navigation}) {
     }
 
     // Database context
-    const { deleteItem, readAllStock, createStock, updateStock, printStock } = useDatabase();
+    const { deleteItem, readAllStock, createStock, updateStock, printStock, checkAdminDelete } = useDatabase();
 
     // Group List
     const [stocks, setStocks] = useState([]);
@@ -79,11 +80,13 @@ export default function NewStock({navigation}) {
 
     // Deleting item function
     const handleDeleteItem = (itemId) => {
-        // Elimina el elemento de la base de datos utilizando deleteItem
-        deleteItem('Product', itemId);
-        
-        // Actualiza el estado groups excluyendo el elemento eliminado
-        setStocks(stocks => stocks.filter(item => item.id !== itemId));
+        checkAdminDelete(itemId, ()=> {
+            // Elimina el elemento de la base de datos utilizando deleteItem
+            deleteItem('Product', itemId);
+            
+            // Actualiza el estado groups excluyendo el elemento eliminado
+            setStocks(stocks => stocks.filter(item => item.id !== itemId));
+        }, showDeleteNotAllowedAlert)
     };
 
     useEffect(() => {
@@ -91,14 +94,22 @@ export default function NewStock({navigation}) {
         readAllStock(setStocks)
     }, []);
 
+    const showDeleteNotAllowedAlert = () => {
+        Alert.alert(
+          'No se puede eliminar el producto',
+          'Aparece en el historial de compras.',
+          [{ text: 'OK', onPress: () => console.log('Alerta cerrada') }]
+        );
+      };
+
     return (
         <View style = {[styles.container,containerExtra]}>
             <ModalStock isVisible={isModalVisible} stock={focusStock} closeModal={closeModal} onCreate={handleAddItem} onEdit={handleEditItem}/>
 
             <View style={styles.hContainer}>
-                <SmallButton title={"Back"} onPress={() => navigation.goBack()} backgroundColor={'white'}></SmallButton>
+                <ImgButton name={'backspace'} onPress={() => navigation.goBack()} backgroundColor={'white'}></ImgButton>
                 <Text> Stock </Text>
-                <SmallButton title={"New\nStock"} onPress={() => openModal(emptyStock)} backgroundColor={'#75F4F4'}></SmallButton>
+                <ImgButton name={'plus'} onPress={() => openModal(emptyStock)} backgroundColor={'#75F4F4'}></ImgButton>
             </View>
             <FlatList
                 contentContainerStyle={styles.productContainer}
@@ -124,7 +135,8 @@ const styles = StyleSheet.create({
       marginLeft:30,
       borderRadius:20,
       borderWidth:5,
-      borderColor: '#d19ba4'
+      borderColor: '#d19ba4',
+      backgroundColor: '#FED8B1'
     },
     productContainer: {
       padding:20,
